FontService: fix potential NPE
This was noticed when trying to use apps that can "freeze" other apps
and "unfreeze" them when needed.

On "unfreezing", it seems that the FontService will try to determine if
the new package is a font provider, while the app is still frozen and
will result in a NPE in FontService. This happened when I try to use the
Island app (with Greenify integration) on the latest build of OmniROM.

This patch simply returns false when package is not found, since
packages that are frozen cannot be used as a system-wide font provider
anyway.

Change-Id: Idf1e44ae72a67bbd433959f30b6f96cb42243036